feat: remove /pages
