+++ src/app/layout.tsx
@@ -3,6 +3,8 @@ import type { Metadata } from "next";
 import { Lato } from "next/font/google";
 import CssBaseline from "@mui/material/CssBaseline";
 import { ThemeProvider } from "@mui/material";
+import Header from "./components/Header";
+import Footer from "./components/Footer";
 
 const lato = Lato({ weight: "400", subsets: ["latin"] });
 // const theme = createTheme();
@@ -23,7 +25,11 @@ export default function RootLayout({
     <html lang="en">
       <CssBaseline />
       {/* <ThemeProvider theme={theme}> */}
-      <body className={lato.className}>{children}</body>
+      <body className={lato.className}>
+        <Header />
+        {children}
+        <Footer />
+      </body>
       {/* </ThemeProvider> */}
     </html>
   );
